Remove outdated function
